fix pixel-accessor integer overflow on large frames (GHSA-ccfw-29x7-rrx3)
The de265_image::get_image_plane_at_pos*() accessors computed the pixel
offset `xpos + ypos*stride` in signed 32-bit int. For frames where
width*height exceeds INT32_MAX (e.g. 46344x46344, both within the existing
<=65535 limit), ypos*stride wraps negative and the returned pointer lands
~2 GB before the buffer -> heap out-of-bounds read/write. This is independent
of the earlier allocation fix (GHSA-vv8h-932h-7r86), which only widened the
buffer size, not the offset arithmetic.

Widen the image stride to ptrdiff_t (members, get_image_stride/get_luma_stride/
get_chroma_stride, set_image_plane) so all pixel-offset math is evaluated in
64 bits. The public C API (de265_get/set_image_plane) keeps int for ABI; the
narrowing at that boundary is safe (stride <= ~65552).

The same 32-bit-stride / absolute-coordinate pattern existed in several scalar
paths that the accessors don't cover; widen those too:
  - sao.cc: SAO offset indexing, and the inputCopy allocation+memcpy size
    (an int size_t under-allocation -> heap write)
  - motion.cc: mc_luma/mc_chroma reference-plane access (ref_stride, src_stride)
  - intrapred.h: intra reference-border fill
  - sei.cc: decoded-picture-hash MD5/CRC/checksum
  - image-io.cc: YUV plane read and write
The deblock kernels, SIMD layer (acceleration.h) and per-block code already
use ptrdiff_t or only small block-local offsets and need no change.

Also store pic_width/height_in_luma_samples as uint16_t (the value is
provably <= MAX_PICTURE_WIDTH = 65535 after validation) so PicSizeInSamplesY
needs only a single uint32_t cast instead of int*int (which was signed-overflow
UB on the same trigger); add static_asserts tying the limits to the storage.

Verified: girlshy bit-exact single- and multi-threaded; ASan+UBSan PoC against
the get_image_plane_at_pos sink crashes pre-fix (UBSan signed-overflow at the
accessor + ASan wild write) and is clean post-fix.

+++ libde265/image-io.cc
@@ -76,7 +76,7 @@ de265_image* ImageSource_YUV::read_next_image()
   // --- load image ---
 
   uint8_t* p;
-  int stride;
+  ptrdiff_t stride;
 
   p = img->get_image_plane(0);  stride = img->get_image_stride(0);
   for (uint32_t y=0;y<height;y++) {
@@ -174,7 +174,7 @@ void ImageSink_YUV::send_image(const de265_image* img)
   // --- write image ---
 
   const uint8_t* p;
-  int stride;
+  ptrdiff_t stride;
 
   int width = img->get_width();
   int height= img->get_height();
